requirements.txt
numpy
scikit-learn
sentence-transformers
google-generativeai
regex